Make typing heatmap last much longer.
This first version increases the value more slowly. From my
understanding of the code the "heat" value is between 0 and 255. This
modification increases the heat by one at every stroke.

The decay of the heat has been increased to decrease by one globally
every 113 seconds. This means that after 8 hours of inactivity all the
keys will have "cooled" down to 0.